Remove network gameplay mode
