+++ README.md
@@ -1,8 +1,10 @@
 # wpf-chess
 
-The simple chess simulator written in C#/WPF. 
-Allows to play with computer or with a friend by network.
+The simple chess simulator written in C#/WPF.
 
 ## Plans
-- practice cross-platform mobile development (Xamarin?)
+
+- practice cross-platform mobile development
 - create Windows Universal App
+- network mode using signalR or sockets or gRPC
+- wixtoolset installer
