--- src/WpfChess.InputLib/Exceptions/NetworkException.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Runtime.Serialization;
-
-namespace WpfChess.InputLib.Exceptions
-{
-    public class NetworkException : InputException
-    {
-        public NetworkException()
-        {
-        }
-
-        public NetworkException(string message) : base(message)
-        {
-        }
-
-        public NetworkException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        protected NetworkException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
-    }
-}
