--- src/WpfChess.InputLib/Services/IDataGet.cs
@@ -1,14 +1,0 @@
-﻿// using System.ServiceModel;
-
-// namespace WpfChess.InputLib.Services
-// {
-//     [ServiceContract]
-//     public interface IDataGet
-//     {
-//         [OperationContract]
-//         bool GetData(Data from, Data to);
-
-//         [OperationContract]
-//         bool IsAvailable();
-//     }
-// }
