+++ src/WpfChess.WpfPresentation/Diagrams/InputLibClassDiagram.cd
@@ -8,10 +8,6 @@
     <Position X="6.75" Y="2.75" Width="1.5" />
     <TypeIdentifier />
   </Class>
-  <Class Name="WpfChess.InputLib.Exceptions.NetworkException" Collapsed="true">
-    <Position X="6.75" Y="4.25" Width="1.5" />
-    <TypeIdentifier />
-  </Class>
   <Class Name="WpfChess.InputLib.Input.AiInputController">
     <Position X="0.75" Y="0.5" Width="1.5" />
     <TypeIdentifier />
@@ -19,11 +15,6 @@
   <Class Name="WpfChess.InputLib.Input.InputController">
     <Position X="3" Y="0.5" Width="1.5" />
     <TypeIdentifier />
-  </Class>
-  <Class Name="WpfChess.InputLib.Input.NetworkInputController">
-    <Position X="5" Y="0.5" Width="1.5" />
-    <TypeIdentifier />
-    <Lollipop Position="0.2" />
   </Class>
   <Class Name="WpfChess.InputLib.Input.OnePcInputController">
     <Position X="0.75" Y="4" Width="1.5" />
