+++ src/WpfChess.WpfPresentation/Model/GameMode.cs
@@ -1,0 +1,7 @@
+namespace WpfChess.WpfPresentation.Model
+{
+    public enum GameMode
+    {
+        OnePcMode, AiMode
+    }
+}
