+++ src/WpfChess.WpfPresentation/View/GamePreloaderView.xaml
@@ -3,43 +3,17 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="Menu" Height="328" Width="370" Background="#FFB49C5B">
     <Window.Resources>
-        <Style x:Key="TextBoxStyle" TargetType="{x:Type TextBox}">
-            <Setter Property="Background">
-                <Setter.Value>
-                    <LinearGradientBrush EndPoint="1,0.5" StartPoint="0,0.5">
-                        <GradientStop Color="#FFBAA2A2" Offset="0.238" />
-                        <GradientStop Color="White" Offset="1" />
-                        <GradientStop Color="#FFDCD0D0" Offset="0.994" />
-                    </LinearGradientBrush>
-                </Setter.Value>
-            </Setter>
+        <Style x:Key="CheckBoxStyle" TargetType="{x:Type CheckBox}">
+            <Setter Property="Background" Value="#FFB78888" />
             <Setter Property="BorderBrush" Value="Brown" />
-            <Setter Property="Foreground" Value="#FF891313" />
-            <Setter Property="Height" Value="23" />
-            <Setter Property="Width" Value="180" />
-            <Setter Property="Margin" Value="0,5,0,0" />
+            <Setter Property="Foreground" Value="#D18F3446" />
+            <Setter Property="Height" Value="16" />
+            <Setter Property="Margin" Value="139,30,0,0" />
         </Style>
-    </Window.Resources>
-    
-    <Grid Height="Auto">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="152*" />
-            <ColumnDefinition Width="196*" />
-        </Grid.ColumnDefinitions>
-
-        <Button Content="One PC mode" Height="56" HorizontalAlignment="Left" VerticalAlignment="Top" Width="127" Style="{StaticResource ButtonStyle}" Margin="12,52,0,0" Command="{Binding Path=GameButtonClickCommand}" CommandParameter="OnePcMode" />
-        <Button Content="AI mode" Height="56" HorizontalAlignment="Left" Margin="10,153,0,0" VerticalAlignment="Top" Width="127" Style="{StaticResource ButtonStyle}" Command="{Binding Path=GameButtonClickCommand}" CommandParameter="AiMode" />
-        
-        <StackPanel Height="261" HorizontalAlignment="Left" Margin="8,16,0,0" VerticalAlignment="Top" Width="188" Grid.Column="1">
-            <CheckBox IsChecked="{Binding Path=IsFirst}" Content="I am white" Height="16" Foreground="#D18F3446" Margin="0,0,0,10" Background="#FFB78888" Padding="10,0,0,0" BorderBrush="Brown" />
-            <TextBlock Text="Your url: //" Foreground="#D18F3446" FontStyle="Normal" FontWeight="ExtraBlack" />
-            <TextBox Text="{Binding Path=ServiceUrl}" Style="{StaticResource TextBoxStyle}" />
-            <TextBlock Text="Game location:" Foreground="#D18F3446" FontStyle="Normal" FontWeight="ExtraBlack" />
-            <TextBox Text="{Binding Path=AdUrl}" Style="{StaticResource TextBoxStyle}" />
-            <TextBlock Text="Second player url-location: //" Foreground="#D18F3446" FontStyle="Normal" FontWeight="ExtraBlack" />
-            <TextBox Text="{Binding Path=EndpointAddress}" Style="{StaticResource TextBoxStyle}" />
-            <Button Content="Network mode" Margin="10,20,10,0" Style="{StaticResource ButtonStyle}" 
-                    Height="56" Command="{Binding Path=GameButtonClickCommand}" CommandParameter="NetworkMode" />
-        </StackPanel>
-    </Grid>
+    </Window.Resources>    
+    <StackPanel>
+        <CheckBox IsChecked="{Binding Path=IsFirst}" Content="I am white" Style="{StaticResource CheckBoxStyle}" />
+        <Button Content="One PC mode" Height="56" Width="127" Margin="12,20,0,0" Style="{StaticResource ButtonStyle}" Command="{Binding Path=GameButtonClickCommand}" CommandParameter="OnePcMode" />
+        <Button Content="AI mode" Height="56" Width="127" Margin="12,30,0,0" Style="{StaticResource ButtonStyle}" Command="{Binding Path=GameButtonClickCommand}" CommandParameter="AiMode" />
+    </StackPanel>
 </Window>
